+++ internal/models/models_generate.go
@@ -49,6 +49,7 @@ func main() {
 	generator.GenerateUpdate(new(models.DiscordConfiguration))
 	generator.GenerateUpdate(new(models.TelegramConfiguration))
 	generator.GenerateUpdate(new(models.CloudConfiguration))
+	generator.GenerateUpdate(new(models.UpdateConfiguration))
 	generator.GenerateUpdate(new(models.SecretConfiguration))
 	generator.GenerateUpdate(new(models.ToolPolicyConfiguration))
 
@@ -91,6 +92,7 @@ func main() {
 	generator.GenerateGetters(new(models.DiscordConfiguration))
 	generator.GenerateGetters(new(models.TelegramConfiguration))
 	generator.GenerateGetters(new(models.CloudConfiguration))
+	generator.GenerateGetters(new(models.UpdateConfiguration))
 	generator.GenerateGetters(new(models.SecretConfiguration))
 	generator.GenerateGetters(new(models.ToolPolicyConfiguration))